Add sections on Persona, Self-Encryption and Proxy
I have included sections that discuss the above topics. Again this is
all very much subject to re-drafting but with this commit my section on
explaining the SAFE Network has all relevant sections. I may decide to
expand on some of the topics and add in diagrams to help clarity.