Start section on introducing the SAFE Network
This commit brings in the beginings of trying to explain, "Why the SAFE
Network exists". This is an important question to try and answer in my
dissertation. By comparing it to other technologies, it brings up
discussion on the benefits (and drawbacks) of the network.

I have spent a great deal of time contrasting the SAFE Network with
BitTorrent. Not only is this a valid comparison, it also helps to ease
the reader into the concepts behind the SAFE Network and how it
operates.

I have also discussed the models client-server/peer-to-peer and how
they contrast/interact with the SAFE Network. Explaining how the
'serverless architecure' model is a good fit for the SAFE Network.