Structural changes and corrections to Chapter 3
I have reduced the size of the diagrams so they don't takeup the whole
page. This in in addition to increasing the clarity of one of them.

I have started going over this section, making a lot of structural
changes which I think are a big improvement. I need to write more so
this chapter will end up larger than originally intended. I have
included the empty sections to indicate my intended structure.