Made big improvements and changes
A number of changes have been made in this commit but overall the report
is far closer to being finished now. Notably the discussion on datamaps,
self-encryption, safecoin and farming has been added. Maths could be
used to further expand and explain some of these points. I am far
happier with the structure and chapter 3 after this commit.