Fix issues with MD and ID explanations
There were some technical inaccuracies when it came to explaining
Immutable and Mutable Data Structures. This has now been fixed so the
dissertation contains what I now feel to be a very solid and accurate
technical explanation of them.

I have added a specific section on MD's which is really sweeeeeet. It
explains how websties work and how you could have an email ``indbox'' on
the network.